+++ TowerDefenceGame/TowerDefenceGame.vcxproj
@@ -123,6 +123,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Button.cpp" />
     <ClCompile Include="Enemy.cpp" />
     <ClCompile Include="Mouse.cpp" />
     <ClCompile Include="Terrain.cpp" />
